labs/python-fastapi-websocket/requirements.txt
# FastAPI and WebSocket support
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1

# OpenTelemetry Core
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# OpenTelemetry Exporters (OTLP for Splunk O11y Cloud)
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-logging==0.50b0

# OpenTelemetry Metrics
opentelemetry-semantic-conventions==0.50b0

# Jinja2 for HTML templates
jinja2==3.1.4
